refactor: change component name to TopicTree
